Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float number;
-    scanf("%f", &number);
+    int number;
+    scanf("%d", &number);
     int started = 0;
     for(int i = 31; i >= 0; i--) {
         int bit = (number>>i) & 1;
